chore(repo-split): stand up root guard harness for opinionated-layer guards (Track A)
Relocates the ~14 opinionated-layer guard tests that lived inside
modules/loop-pipeline/tests/ (the repo's only CI-covered test tree) purely
because that was the only place CI ran root-reaching assertions. Per
DESIGN-repo-split.md Phase 1's riskiest item #2: when loop-pipeline is
eventually extracted to the runner repo, these guards must already be
running from repo-root tests/, or they silently stop existing with CI
green.

- Moved wholesale (no engine coupling): test_quality_protocol_guard.py,
  test_doc_consistency.py (majority), test_explainer_doc_guard.py
  (majority), test_orchestrator_source_pin_guard.py.
- Split (engine half stays, opinionated half moves): test_authoring_layer_gates.py,
  test_objective_layer_gates.py, test_drift_review_gate.py,
  test_engine_semantics_doc_guard.py, test_doc_consistency.py (2 functions),
  test_explainer_doc_guard.py (4 functions) -- the residual halves assert
  live engine behavior (parse_dot/lint/PipelineEngine) and stay in
  modules/loop-pipeline/tests/, each with a provenance note.
- Left in place (engine/spec-machinery, rides with the runner at
  extraction): test_spec_conformance_matrix.py,
  test_extensions_ledger_integrity.py, and the examples-corpus lint/render
  sweeps that fundamentally require the live linter/parser
  (test_examples_lint_clean.py, test_exemplar_terminals.py,
  test_inert_vocabulary_lint.py, test_topological_lint.py,
  test_dot_render_compliance.py, test_inert_evidence_gate_calibration.py) --
  moving these without an engine dependency would either duplicate the
  linter or reintroduce the exact root-tests-import-loop-pipeline coupling
  this PR exists to avoid.
- Re-anchored repo-root resolution in every moved file
  (Path(__file__).resolve().parent.parent, since tests/ is now one level
  below repo root) instead of the old modules/loop-pipeline-relative
  parents[N]/BUNDLE_ROOT depth.
- Updated docs/QUALITY_PROTOCOL.md's Q-300 file-name references to the
  guards' new tests/ location.
- Added "Opinionated Guards (repo root)" CI job (installs only pytest, no
  module deps) and wired it into the ci-gate aggregate's needs list.

Zero engine code changes. Negative controls verified: mutating the Q-307
decision-matrix text in docs/VISION.md fails
test_q307_decision_matrix_reads_identically_on_both_pages from root;
flipping an orchestrator source: to a relative path in
bundles/attractor-agent.yaml fails the OSP guard from root. Both restored
green.
